Prompts for different contrib types, create settings yml, styling changes
