--- console/ai_validation2.php
@@ -1,455 +1,0 @@
-<?php
-/** 
- * Note: This file was generated by Copilot as a 'cleaned up' version of ai_validation.php
- */
-
-namespace phpbb\oberon\console;
-
-use Symfony\Component\Console\Command\Command;
-use Symfony\Component\Console\Input\InputInterface;
-use Symfony\Component\Console\Output\OutputInterface;
-use Symfony\Component\Console\Input\InputArgument;
-
-class ai_validation2 extends Command
-{
-    /* @var \phpbb\db\driver\driver_interface $db */
-    protected $db;
-
-    /* @var \phpbb\oberon\manager\manager $manager */
-    protected $manager;
-
-    public function __construct(\phpbb\db\driver\driver_interface $db, \phpbb\oberon\manager\manager $manager)
-    {
-        $this->db = $db;
-        $this->manager = $manager;
-        parent::__construct();
-    }
-
-    protected function configure()
-    {
-        // Run the command with: php bin/phpbbcli.php custdb:ai_validation2 2
-        $this
-            ->setName('custdb:ai_validation2')
-            ->setDescription('Validate a Customisation DB contribution revision using OpenAI and return a structured report')
-            ->setHelp('This command runs a structured OpenAI validation against a contribution revision in the queue.')
-            ->addArgument(
-                'queue_id', // Argument name
-                InputArgument::REQUIRED, // Required argument
-                'The ID of the queue item to process'
-            );
-    }
-
-    protected function execute(InputInterface $input, OutputInterface $output)
-    {
-        $queue_id = (int) $input->getArgument('queue_id');
-        $output->writeln('<info>[execute] STARTING AI VALIDATION (v2)</info>');
-
-        $result = $this->process_queue($queue_id, $output);
-
-        if (!$result)
-        {
-            $output->writeln('<error>[execute] AI validation failed.</error>');
-            return Command::FAILURE;
-        }
-
-        $output->writeln('<info>[execute] FINISHED AI VALIDATION (v2)</info>');
-        return Command::SUCCESS;
-    }
-
-    private function process_queue(int $queue_id, OutputInterface $output): bool
-    {
-        $output->writeln("<info>Loading queue item {$queue_id}...</info>");
-
-        $queue_item = $this->manager->find_contribution_revision_for_queue_id($queue_id);
-
-        // Change the internal queue status to "awaiting AI validation"
-        $this->manager->update_internal_queue_status($queue_id, $this->manager::INTERNAL_STATUS_AWAITING_AI_VALIDATION);
-
-        if (empty($queue_item['revision']))
-        {
-            $output->writeln('<error>Unable to find revision for queue item.</error>');
-            return false;
-        }
-
-        $revision = $queue_item['revision'];
-        $contribution = $queue_item['contribution'];
-
-        $zip_path = $this->manager::FILE_UPLOAD_LOCATION . $revision['revision_attachment'];
-        if (!file_exists($zip_path))
-        {
-            $output->writeln("<error>Zip file not found: {$zip_path}</error>");
-            return false;
-        }
-
-        $output->writeln("<info>Extracting ZIP: {$zip_path}</info>");
-        $extract_dir = $this->extract_zip($zip_path, $output);
-        if (!$extract_dir)
-        {
-            $output->writeln('<error>Failed to extract zip.</error>');
-            return false;
-        }
-
-        try
-        {
-            $files = $this->collect_files($extract_dir, $output);
-            $manifest = $this->build_manifest($files);
-            $messages = $this->build_openai_messages($contribution, $revision, $manifest, $files);
-            $response = $this->call_openai($messages, $output);
-
-            // Replace characters which could cause the JSON decoding to crash
-            $response = preg_replace('/[\x00-\x1F\x7F]/u', '', $response);
-
-            $output->writeln('<info>OpenAI response (raw):</info>');
-            $output->writeln($response);
-
-            $decoded = json_decode($response, true);
-
-            if (json_last_error() === JSON_ERROR_NONE && isset($decoded['outcome']))
-            {
-                $output->writeln('<info>---- Parsed Validation Report ----</info>');
-                $output->writeln(sprintf('<info>Name:</info> %s', $decoded['name'] ?? '')); 
-                $output->writeln(sprintf('<info>Outcome:</info> %s', $decoded['outcome'] ?? '')); 
-                $output->writeln(sprintf('<info>Confidence:</info> %s', $decoded['confidence'] ?? '')); 
-                $output->writeln('<info>Report:</info>');
-                
-                $report = $decoded['report'] ?? '';
-
-                // Convert the validation report to human-readable format
-                $report = $this->format_validation_report($report);
-                
-                $output->writeln($report);
-
-                // Store the validation report in a new post/topic
-                $topic_id = $this->manager->store_validation_report(
-                    (int) $contribution['contribution_id'],
-                    $report,
-                    $decoded['outcome'] ?? 'unknown',
-                    (int) ($decoded['confidence'] ?? 0)
-                );
-
-                if ($topic_id)
-                {
-                    $output->writeln("<info>Stored validation report in topic id {$topic_id}.</info>");
-                }
-                else
-                {
-                    $output->writeln('<comment>Failed to store validation report in forum topic.</comment>');
-                }
-
-                // Change the internal queue status to "completed AI validation"
-                $this->manager->update_internal_queue_status($queue_id, $this->manager::INTERNAL_STATUS_COMPLETED_AI_VALIDATION);
-            }
-            else
-            {
-                $output->writeln('<comment>Unable to parse response as JSON with expected keys. Please inspect output above.</comment>');
-                $output->writeln('Error code: ' . json_last_error());
-                $output->writeln('Error message: ' . json_last_error_msg());
-            }
-
-            return true;
-        }
-        finally
-        {
-            $output->writeln("<info>Cleaning up extraction folder: {$extract_dir}</info>");
-            $this->delete_directory_recursive($extract_dir);
-        }
-    }
-
-    private function extract_zip(string $zip_path, OutputInterface $output): ?string
-    {
-        $extract_dir = sys_get_temp_dir() . '/phpbb_ai_validation2_' . uniqid();
-        if (!is_dir($extract_dir))
-        {
-            mkdir($extract_dir, 0777, true);
-        }
-
-        $zip = new \ZipArchive();
-        if ($zip->open($zip_path) === TRUE)
-        {
-            $zip->extractTo($extract_dir);
-            $zip->close();
-            return $extract_dir;
-        }
-
-        return null;
-    }
-
-    private function collect_files(string $root_dir, OutputInterface $output): array
-    {
-        $files = [];
-        $iterator = new \RecursiveIteratorIterator(
-            new \RecursiveDirectoryIterator($root_dir, \RecursiveDirectoryIterator::SKIP_DOTS)
-        );
-
-        // Extensions that we treat as text for inclusion.
-        $text_extensions = [
-            'php', 'html', 'htm', 'css', 'js', 'json', 'xml', 'md', 'txt', 'yml', 'yaml', 'csv',
-            'py', 'rb', 'ts', 'c', 'cpp', 'h', 'java', 'go', 'sh', 'ini', 'conf', 'bat', 'ps1',
-        ];
-
-        foreach ($iterator as $file)
-        {
-            if (!$file->isFile())
-            {
-                continue;
-            }
-
-            $rel_path = str_replace($root_dir . DIRECTORY_SEPARATOR, '', $file->getPathname());
-            $extension = strtolower(pathinfo($rel_path, PATHINFO_EXTENSION));
-            $size = $file->getSize();
-
-            if ($size === 0)
-            {
-                $files[$rel_path] = [
-                    'path' => $rel_path,
-                    'size' => 0,
-                    'extension' => $extension,
-                    'content' => '',
-                    'status' => 'empty',
-                ];
-                continue;
-            }
-
-            if (!in_array($extension, $text_extensions, true))
-            {
-                $output->writeln("<comment>Skipping non-text file: {$rel_path}</comment>");
-                $files[$rel_path] = [
-                    'path' => $rel_path,
-                    'size' => $size,
-                    'extension' => $extension,
-                    'content' => null,
-                    'status' => 'binary_skipped',
-                ];
-                continue;
-            }
-
-            $content = file_get_contents($file->getPathname());
-            if ($content === false)
-            {
-                $output->writeln("<comment>Failed to read file: {$rel_path}</comment>");
-                $content = '';
-            }
-
-            // Truncate large files to avoid exceeding API limits
-            $max_chars = 120000;
-            if (strlen($content) > $max_chars)
-            {
-                $content = substr($content, 0, $max_chars);
-                $output->writeln("<comment>Truncated file content for: {$rel_path} (>{$max_chars} chars)</comment>");
-            }
-
-            $files[$rel_path] = [
-                'path' => $rel_path,
-                'size' => $size,
-                'extension' => $extension,
-                'content' => $content,
-                'status' => 'included',
-            ];
-        }
-
-        return $files;
-    }
-
-    private function build_manifest(array $files): array
-    {
-        $manifest = [];
-        foreach ($files as $file)
-        {
-            $manifest[] = [
-                'path' => $file['path'],
-                'size' => $file['size'],
-                'extension' => $file['extension'],
-                'status' => $file['status'],
-            ];
-        }
-        return $manifest;
-    }
-
-    private function build_openai_messages(array $contribution, array $revision, array $manifest, array $files): array
-    {
-        $system_prompt = $this->load_prompt_file('system.prompt.txt');
-        $user_prompt = $this->load_prompt_file('user.prompt.txt');
-
-        $manifest_json = json_encode($manifest, JSON_PRETTY_PRINT);
-        $file_overview = "Manifest (showing path, extension, size, status):\n" . $manifest_json;
-
-        $included_files = [];
-        foreach ($files as $file)
-        {
-            if ($file['status'] !== 'included')
-            {
-                continue;
-            }
-
-            $included_files[] = [
-                'path' => $file['path'],
-                'content' => $file['content'],
-            ];
-        }
-
-        // Assemble a single prompt chunk with file contents.
-        $file_contents_sections = [];
-        foreach ($included_files as $item)
-        {
-            $file_contents_sections[] = "---\nFile: {$item['path']}\n---\n" . $item['content'];
-        }
-
-        $file_contents = implode("\n\n", $file_contents_sections);
-        if (strlen($file_contents) > 250000)
-        {
-            // Trim down if it grows too large
-            $file_contents = substr($file_contents, 0, 250000);
-            $file_contents .= "\n\n... (truncated due to length)";
-        }
-
-        $user_message = "Contribution: {$contribution['contribution_name']}\n" .
-                        "Contribution ID: {$contribution['contribution_id']}\n" .
-                        "Revision ID: {$revision['revision_id']}\n" .
-                        "Revision name: {$revision['revision_name']}\n" .
-                        "Revision version: {$revision['revision_version']}\n" .
-                        "\n" .
-                        "{$user_prompt}\n" .
-                        "\n" .
-                        "{$file_overview}\n" .
-                        "\n" .
-                        "Files:\n" .
-                        "{$file_contents}\n";
-
-        return [
-            ['role' => 'system', 'content' => $system_prompt],
-            ['role' => 'user', 'content' => $user_message],
-        ];
-    }
-
-    private function load_prompt_file(string $file_name): string
-    {
-        $path = __DIR__ . '/' . $file_name;
-        if (!file_exists($path))
-        {
-            return '';
-        }
-        return trim(file_get_contents($path));
-    }
-
-    private function call_openai(array $messages, OutputInterface $output): string
-    {
-        $api_key = $this->get_openai_api_secret();
-        if (!$api_key)
-        {
-            throw new \RuntimeException('OpenAI API key is not set in environment.');
-        }
-
-        $model = getenv('PHPBB_CUSTDB_OPENAI_MODEL') ?: 'gpt-4o-mini';
-
-        $payload = [
-            'model' => $model,
-            'messages' => $messages,
-            'temperature' => 0.0,
-            'max_tokens' => 4000, // this is a larger response; if a lower one is used, the JSON could cut off.
-        ];
-
-        $ch = curl_init('https://api.openai.com/v1/chat/completions');
-        curl_setopt($ch, CURLOPT_RETURNTRANSFER, true);
-        curl_setopt($ch, CURLOPT_POST, true);
-        curl_setopt($ch, CURLOPT_HTTPHEADER, [
-            'Content-Type: application/json',
-            'Authorization: Bearer ' . $api_key,
-        ]);
-        curl_setopt($ch, CURLOPT_POSTFIELDS, json_encode($payload));
-
-        $response = curl_exec($ch);
-        $error = curl_error($ch);
-        $status = curl_getinfo($ch, CURLINFO_HTTP_CODE);
-        curl_close($ch);
-
-        if ($error) {
-            throw new \RuntimeException('OpenAI request failed: ' . $error);
-        }
-
-        if ($status < 200 || $status >= 300) {
-            throw new \RuntimeException('OpenAI returned HTTP ' . $status . ': ' . $response);
-        }
-
-        $data = json_decode($response, true);
-        if (!isset($data['choices'][0]['message']['content'])) {
-            throw new \RuntimeException('Unexpected OpenAI response format: ' . $response);
-        }
-
-        return $data['choices'][0]['message']['content'];
-    }
-
-    private function delete_directory_recursive($dir)
-    {
-        if (!is_dir($dir)) return;
-        $items = new \RecursiveIteratorIterator(
-            new \RecursiveDirectoryIterator($dir, \RecursiveDirectoryIterator::SKIP_DOTS),
-            \RecursiveIteratorIterator::CHILD_FIRST
-        );
-        foreach ($items as $item) 
-        {
-            if ($item->isDir()) 
-            {
-                rmdir($item->getPathname());
-            } 
-            
-            else 
-            {
-                unlink($item->getPathname());
-            }
-        }
-        rmdir($dir);
-    }
-
-    /**
-     * Convert validation report to human-readable BBCode format
-     *
-     * Converts either a JSON array or formatted string report into a structured
-     * forum post with file names, code snippets, and validation tags.
-     *
-     * @param mixed $report The report data (array or string)
-     * @return string Formatted BBCode post content
-     */
-    private function format_validation_report($report): string
-    {
-        if (is_array($report))
-        {
-            $formatted = '';
-            
-            foreach ($report as $item)
-            {
-                $filename = $item['file'] ?? 'Unknown File';
-                $issues = $item['issues'] ?? [];
-                
-                // File heading
-                $formatted .= "\n[b]File: {$filename}[/b]\n\n";
-                
-                if (is_array($issues) && !empty($issues))
-                {
-                    foreach ($issues as $issue)
-                    {
-                        $code = $issue['code'];// ?? 'valinfo';
-
-                        $formatted .= "\n[code]{$code}[/code]\n\n";
-
-                        $message = $issue['message'] ?? '';
-                        $formatted .= "\n{$message}\n\n";
-                    }
-                }
-                else
-                {
-                    // Optional fallback if no issues exist
-                    $formatted .= "[valinfo]\nNo issues found.\n[/valinfo]\n\n";
-                }
-            }
-            
-            return trim($formatted);
-        }
-        
-        return (string) $report;
-    }
-
-    private function get_openai_api_secret(): ?string
-    {
-        return getenv('PHPBB_CUSTDB_TEST_CHATGPT_API_AI_KEY');
-    }
-}
