+++ console/prompts/extension.user.prompt.txt
